moved attachPreviews to a separate file
