+++ apps/web/src/utils/messages.ts
@@ -5,10 +5,14 @@ SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Com
 Please see LICENSE files in the repository root for full details.
 */
 
-import { type MatrixEvent, type IContent, type IMentions, type IEventRelation } from "matrix-js-sdk/src/matrix";
+import { type MatrixEvent, type IContent, type IMentions, type IEventRelation, MatrixClient } from "matrix-js-sdk/src/matrix";
 
 import type EditorModel from "../editor/model";
 import { Type } from "../editor/parts";
+import { MessageComposerUrlPreviewSnapshot } from "@element-hq/web-shared-components";
+import { RoomMessageEventContent } from "../../@types/url-preview";
+import { EncryptedFile } from "matrix-js-sdk/src/types";
+import { uploadFile } from "../ContentMessages";
 
 /**
  * Build the mentions information based on the editor model (and any related events):
@@ -107,3 +111,53 @@ export function attachRelation(content: IContent, relation?: IEventRelation): vo
         };
     }
 }
+
+export async function attachUrlPreviews(
+    mxClient: MatrixClient,
+    roomId: string,
+    urlPreviewSnapshot: MessageComposerUrlPreviewSnapshot,
+    content: RoomMessageEventContent,
+): Promise<void> {
+    if (urlPreviewSnapshot.previews.length) {
+        content["com.beeper.linkpreviews"] = await Promise.all(
+            urlPreviewSnapshot.previews.map(async (preview) => {
+                // upload the files to produce the mxc:// url for the images
+                let imageUploaded: { url?: string; file?: EncryptedFile } | null = null;
+
+                if (preview.image) {
+                    let imageBlob: Blob | null = null;
+
+                    try {
+                        imageBlob = await (await fetch(preview.image.imageFull)).blob();
+                    } catch (e) {
+                        console.error(`Failed to fetch image from ${preview.image.imageFull}`, e);
+                    }
+
+                    if (imageBlob !== null) {
+                        try {
+                            imageUploaded = await uploadFile(mxClient, roomId, imageBlob);
+                        } catch (e) {
+                            console.error("Failed to upload image", e);
+                        }
+                    }
+                }
+
+                return {
+                    "matched_url": preview.link,
+                    // TODO: og:url may be different from the URL requested, but is not present in the response of the url_preview
+                    "og:url": preview.link,
+                    "og:title": preview.title,
+                    "og:description": preview.description,
+                    "og:image": imageUploaded?.url,
+                    // TODO: should we trust the declared size? or should we calculate it ourselves
+                    "og:image:width": preview.image?.width,
+                    "og:image:height": preview.image?.height,
+                    "og:image:type": preview.image?.imageType,
+                    "matrix:image:size": preview.image?.fileSize,
+                    "beeper:image:encryption": imageUploaded?.file,
+                };
+            }),
+        );
+    }
+}
+
